Extract model class
